bugfix: heredoc時 Ctrl+C でプロンプトが2個出ることがある問題 https://github.com/ryo-manba/minishell/issues/47
